# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payment QR Service API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders/{OrderUnit::uuid}": {
            "get": {
                "tags": [
                    "Order Unit"
                ],
                "summary": "Получить заказ по uuid.",
                "operationId": "6f875155dfa4800ea488578fc462b057",
                "parameters": [
                    {
                        "name": "OrderUnit::uuid",
                        "in": "path",
                        "description": "UUID заказа",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный возврат заказа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/OrderUnitResource"
                                                }
                                            ],
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Return Order."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Общая ошибка сервера.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Error server"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Order Unit"
                ],
                "summary": "Обновить данные заказа.",
                "description": "Обновляет информацию о заказе по UUID",
                "operationId": "14a9a4f41dcce93793f26265c94b80c3",
                "parameters": [
                    {
                        "name": "OrderUnit::uuid",
                        "in": "path",
                        "description": "UUID заказа",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "change_price": {
                                                "type": "boolean",
                                                "example": false,
                                                "nullable": true
                                            },
                                            "change_time": {
                                                "type": "boolean",
                                                "example": true,
                                                "nullable": true
                                            },
                                            "order_status": {
                                                "type": "string",
                                                "enum": [
                                                    "Опубликован",
                                                    "Закрыт",
                                                    "Отозван"
                                                ],
                                                "example": "Опубликован",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Обновление заказа по указанным параметрам, успешно.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Update order successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка обновление заказа.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Update order error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Общая ошибка сервера.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Error server"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/get-schem": {
            "post": {
                "tags": [
                    "Order Unit"
                ],
                "summary": "Получить все заказы которые входят в вектор движения Главного заказа",
                "operationId": "6b495ec03837a5b38c46cb99abee5048",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "main_order": {
                                                "description": "ID основного заказа, обязательное поле",
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "550e8400-e29b-41d4-a716-446655440000"
                                            },
                                            "search_distance": {
                                                "description": "Расстояние для поиска, необязательное поле - по стандарту стоит 100 км (Поиска) - (Указывать в Километрах)",
                                                "type": "integer",
                                                "example": 10
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный возврат заказов входящие в область.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/OrderUnitResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Return Orders."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Общая ошибка сервера.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Error server"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "OrderUnitResource": {
                "title": "Json Ресурс Order Unit",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "delivery_start": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "delivery_end": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "address_start_id": {
                        "$ref": "#/components/schemas/addressResource"
                    },
                    "address_end_id": {
                        "$ref": "#/components/schemas/addressResource"
                    },
                    "body_volume": {
                        "type": "number",
                        "format": "float"
                    },
                    "order_total": {
                        "type": "number",
                        "format": "float"
                    },
                    "description": {
                        "type": "string"
                    },
                    "product_type": {
                        "type": "string"
                    },
                    "order_status": {
                        "type": "string"
                    },
                    "user_id": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "organization_id": {
                        "$ref": "#/components/schemas/OrganizationResource"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}